feat: prototype manifest build
DO NOT MERGE

This patch implements a PoC of the manifest build which uses
the instructions as found in the `[build]` section to build
the package exactly as it would be built by manually invoking
the same commands within a `flox activate` shell.

- sets $out in the environment
- invokes the build commands in a subshell (using bash)
- turns the $out directory into a Nix package with all outpath
  references replaced with the real $out
- wraps all `bin/*` commands with `$FLOX_ENV/activate`

Usage: flox-build [ pkgname ]

N.B. this patch utilises the (hitherto-unused) $FLOX_TURBO flag
to invoke the provided command without an intervening shell, and
in the process I tripped across and fixed a bug in the activate
script.