feat: substitute path references for package dependencies in builds
This patch adds the feature of rewriting the build script to substitute
all references to `${pkgname}` within a build script with their storePath
values as derived from `readline result-pkgname` prior to kicking off the
builds. With this feature we can automatically infer the dependency graph
and configure builds to depend upon each other.