feat(watchdog): add signal listener
The watchdog now listens for signals. Each task also now returns an
`Action` which is either `Terminate` or `Cleanup` since some signals
can trigger shutdown (SIGINT, etc) and some signals can trigger cleanup
(SIGUSR1). Similarly, the termination task also returns `Cleanup` when
it's returning because it detected a termination, or `Terminate` if it
observed that the shutdown flag was set.

The trickiest part of this was realizing that tasks spawned with
`tokio::task::spawn_blocking` *cannot be aborted* via the `abort()`
method on the task's `JoinHandle` (which you *can* do for tasks spawned
via `tokio::spawn`). The reason behind this is that `spawn_blocking`
tasks are not async (they're just threads), and the `abort` method
attempts to cancel a task at the next `.await` point that it hits. Since
a thread is not async, it never hits an `.await` point, and therefore
cannot be aborted this way. This is why we need the shutdown flag, so
that all tasks, regardless of how they were spawned, can receive some
kind of notification that it's time to shut down.

I came to this realization after noticing that even after receiving a
cleanup signal, the application would hang at the end of `main`,
presumably because something was waiting for the macOS termination
listener, which is the *only* `spawn_blocking` task. For reasoning about
why macOS is a `spawn_blocking` task, see the module-level documentation
in `listen.rs`. In short, waiting for a termination on macOS was
blocking before using the shutdown flag, whereas on Linux it's async
(you're waiting for a signal). The macOS implementation was changed from
a blocking call to a poll-based method so that the shutdown flag could
be periodically checked.