feat(registry): Track activation PIDs
This will be used by the watchdog `klaus` to track when the last
activation of an environment exits and we should clean up any running
services.

Stale activations are removed when updating the registry on disk to
catch watchdog processes that may have died without cleaning up after
themselves.

The new field needs to use `default` in order to read registry files
from existing Flox versions where the field isn't present.

Some assumptions in the implementation:

- environments are already registered by the time we're activating or
  starting the watchdog; an error will be returned if not
- an activation will only be registered and deregistered once;
  subsequent calls are noops
- we can't record activations for remote environments because they
  aren't currently registered in the environment registry; that's not
  currently an issue for service management because we don't support
  services for remote environments at the moment
- testing focuses on unit testing the inner parts because the file
  handling part is relatively simple, proven, and more fiddly to
  co-ordinate in a series of tests