refactor(watchdog): separate components for tests
This refactors the body of `main` into a function that takes tasks as
arguments so that unit tests can provide their own dummy implementations
of each task. It also adds a few unit tests.

The `MockListener` struct is used to mock out signals delivered to the
application, and can only be constructed in tests.